[KK][GT] Cache temporary arrays in ChaControl.UpdateVisible
This is another attempt at 3124e14, which got reverted.

Original commit message:
> This patch caches the 1D and 2D temporary arrays created in
> UpdateVisible.
>
> The transpiler is a little blunt, replacing all newarr instructions,
> so I included a check to ensure that we patch exactly 45 places.
> This should make it a little more robust against other plugins
> transpiling the same method.

The previous patch broke ClothesToAccessories when UpdateVisible
contained an unexpected number of array allocations. This happened
because the transpiler threw an exception, which, depending
on the patch order, could prevent the transpiler defined by
ClothesToAccessories (or any other plugin that patches the method)
from running.

This version avoids the issue by logging an error instead of
throwing an exception.

This version also removes the use of CodeMatcher.SetAndAdvance
because it looks like it could destructively modify input code.