LoadFileLimitedFix (#68)
Makes LoadFileLimited overwrite the charaFileName if new character parameters were loaded.
This is useful if plugins what to know what character is currently loaded. Players are most likely to think of a character as currently loaded depending on the parameters (name, birthday, etc), so the fix will not change the charaFileName unless those parameters were replaced when loading a character in maker.