[DynamicBoneFix] Fix flexible gimmicks. (#81)
Fixes the "bug" breaking some scenes mentioned in #76 by fixing the premature end of the dynamic bone chain as explained in my comment.

This PR intercepts the AddObjectItem.Load process and adds the "N_setuzoku" bone to the m_Exlcusions list and clears the m_notRolls list. Like that the dynamic bone chain correctly ends on the gimmick again and does not continue into the item parented to it.